+++ packages/snippets/src/ant-design-x/src/sender.vue
@@ -8,10 +8,11 @@
 
 <script lang="ts" setup>
   import type { SenderProps } from '@ant-design/x';
-  import { Sender as XSender } from '@ant-design/x';
   import { omit, transformKeys } from 'mixte';
   import { applyPureReactInVue, injectSyncUpdateForPureReactInVue } from 'veaury';
-  import { computed, useAttrs } from 'vue';
+  import { computed, onBeforeUpdate, ref, useAttrs } from 'vue';
+  import WrappedSender from './components-react/sender';
+  import { senderSlots, type SenderSlots } from './types';
 
   interface Props extends /* @vue-ignore */ Partial<
     Omit<SenderProps, 'value' | 'allowSpeech' | 'disabled' | 'loading' | 'readOnly'>
@@ -29,22 +30,38 @@
   defineOptions({ inheritAttrs: false });
 
   const props = defineProps<Props>();
+  const slots = defineSlots<SenderSlots>();
   const attrs = useAttrs() as SenderProps;
 
   const value = defineModel<string>('modelValue', {
     default: '',
   });
 
-  injectSyncUpdateForPureReactInVue(XSender, {
+  injectSyncUpdateForPureReactInVue(WrappedSender, {
     onChange: (value: string) => ({ value }),
   });
 
-  const Sender = applyPureReactInVue(XSender);
+  const Sender = applyPureReactInVue(WrappedSender);
+
+  /**
+   * 响应式无法监听到组件插槽变化, 需要强制使响应式重新计算
+   * @see https://github.com/vuejs/core/issues/11227
+   */
+  const hasSlots = ref(
+    senderSlots.map(key => !!slots[key]).join(','),
+  );
+
+  onBeforeUpdate(() => {
+    hasSlots.value = senderSlots.map(key => !!slots[key]).join(',');
+  });
 
   const senderProps = computed(() => {
+    hasSlots.value; // eslint-disable-line ts/no-unused-expressions
+
     return {
       ...props,
       ...transformKeys(omit(attrs, ['value', 'onChange'])),
+      ...slots,
     };
   });